src/sku-cleanup-tool/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sku-cleanup-tool"
version = "1.0.0"
description = "Automated cleanup tool for obsolete Amazon FBA inventory"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "boto3>=1.28.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
sku-cleanup = "sku_cleanup_tool.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["sku_cleanup_tool*"]

[tool.setuptools.package-data]
sku_cleanup_tool = ["config/*.txt", "config/*.ini", "config/*.example"]
